+++ packages/app/src/response.ts
@@ -3,7 +3,7 @@ import * as cookie from '@tinyhttp/cookie'
 import { Request } from './request'
 import { App, TemplateEngineOptions } from './app'
 
-import type { ReadStreamOptions } from '@tinyhttp/res'
+import type { ReadStreamOptions, FormatProps } from '@tinyhttp/res'
 
 export const renderTemplate = (_req: Request, res: Response, app: App) => (file: string, data?: Record<string, any>, options?: TemplateEngineOptions): Response => {
   app.render(
@@ -34,4 +34,6 @@ export interface Response extends ServerResponse {
   links(links: { [key: string]: string }): Response
   render(file: string, data?: Record<string, any>, options?: TemplateEngineOptions): Response
   vary(field: string): Response
+  format(obj: FormatProps): Response
+  redirect(url: string, status?: number): Response
 }