+++ packages/res/src/index.ts
@@ -1,3 +1,6 @@
 export * from '@tinyhttp/send'
 export * from './cookie'
 export * from './headers'
+export type { FormatProps } from './format'
+export { formatResponse } from './format'
+export * from './redirect'