+++ packages/ip-filter/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @tinyhttp/ip-filter
 
+## 0.1.7
+
+### Patch Changes
+
+- @tinyhttp/app@0.3.7
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.3.6
+
 ## 0.1.5
 
 ### Patch Changes